Changes in code; -
line 48 whre a primitive variable is used to compare the variable stored in memory. 
lines 73 - 77 were added to get the location of the tomato's which is used on all greeps which is in the "act" method.
lines 109 - 111 were added to save the location if tomato's were found in the "checkfood" method

These bits of code were taken from the SimpleGreep.